refactor: refactor Jinja generator module path
